Add if statement

+++ admin_scripts/ubuntu/ubuntu_16.04/layout/remove_new_release_message.sh
@@ -4,7 +4,10 @@
 release_upgrades_file=/etc/update-manager/release-upgrades
 
 # Simple backup
-cp $release_upgrades_file $release_upgrades_file.org
+if [ ! -f $release_upgrades_file.org ]
+then
+    cp $release_upgrades_file $release_upgrades_file.org
+fi
 
 # Replace Prompt with never value
-sed -i 's/Prompt=.*/Prompt=never/' release_upgrades_file
+sed -i 's/Prompt=.*/Prompt=never/' $release_upgrades_file
